Alertas para confirmación de edición y eliminación de post

+++ src/view/home.js
@@ -6,60 +6,60 @@ import {
 export default () => {
   // CREACIÓN DEL TEMPLATE
   const viewHome = `
-    <section class="container-profile">
-      <div class="profile-card">
-        <div class="img-card-profile">
-          <img src="img/perfilwhite.png" alt="imagen de perfil">
+    <section class='container-profile'>
+      <div class='profile-card'>
+        <div class='img-card-profile'>
+          <img src='img/perfilwhite.png' alt='imagen de perfil'>
         </div>
-        <div class="user-name">
+        <div class='user-name'>
           <h5>${auth.currentUser.email}</h5>
         </div>
-        <div class="user-info">
-          <div class="about-user">
-            <span class="about-me">Mis juegos favoritos son...</span>
+        <div class='user-info'>
+          <div class='about-user'>
+            <span class='about-me'>Mis juegos favoritos son...</span>
           </div>
-          <div class="more-info">
-            <span class"more-about-me">Interacción</span>
+          <div class='more-info'>
+            <span class'more-about-me'>Interacción</span>
           </div>
         </div>
       </div>
-      <div class="info-dev">
+      <div class='info-dev'>
         <span>Desarrolado por</span>
-        <ul class="menu-info-dev">
+        <ul class='menu-info-dev'>
           <li>
-            <a href="https://github.com/OryChRamirez" title="Orayma Chacón"><i class="bi bi-github"></i></a>
+            <a href='https://github.com/OryChRamirez' title='Orayma Chacón'><i class='bi bi-github'></i></a>
           </li>
           <li>
-            <a href="https://github.com/Bellasacc" title="Bella Aguirre"><i class="bi bi-github"></i></a>
+            <a href='https://github.com/Bellasacc' title='Bella Aguirre'><i class='bi bi-github'></i></a>
           </li>
           <li>
-            <a href="https://github.com/daniandp" title="Daniela Andrade"><i class="bi bi-github"></i></a>
+            <a href='https://github.com/daniandp' title='Daniela Andrade'><i class='bi bi-github'></i></a>
           </li>
         </ul>
       </div>
     </section>
-    <section class="container-main">
-      <div class="container-input-post">
-        <div class="cont-descriptions-post">
-          <span class="span-text"> Crea una publicación</span>
-          <span role="textbox" contenteditable id="create-post" placeholder="¿Qué quieres compartir, gamer?"></span>
+    <section class='container-main'>
+      <div class='container-input-post'>
+        <div class='cont-descriptions-post'>
+          <span class='span-text'> Crea una publicación</span>
+          <span role='textbox' contenteditable id='create-post' placeholder='¿Qué quieres compartir, gamer?'></span>
         </div>
-        <div class="container-share-btn">
-          <div class="share-img" >
-            <i class="bi bi-image bi-size"></i>
-            <span class="span-text"> Agregar imagen</span>
+        <div class='container-share-btn'>
+          <div class='share-img' >
+            <i class='bi bi-image bi-size'></i>
+            <span class='span-text'> Agregar imagen</span>
           </div>
           <div>
-          <input id="check-private" type="checkbox">
+          <input id='check-private' type='checkbox'>
           Privado
           </div>
-          <div class="share-post">
-            <button type="button" id="btn-share-post">Publicar</button>
+          <div class='share-post'>
+            <button type='button' id='btn-share-post'>Publicar</button>
           </div>
         </div>
       </div>
-      <h3 class="text-publications">PUBLICACIONES</h3>
-      <div class="container-post"></div>
+      <h3 class='text-publications'>PUBLICACIONES</h3>
+      <div class='container-post'></div>
     </section>`;
 
   // CREANDO NODO SECTION
@@ -96,42 +96,42 @@ export default () => {
         const divPostPublicated = document.createElement('div');
         divPostPublicated.setAttribute('class', 'container-publicated');
         divPostPublicated.innerHTML = `
-        <div class="post-publicated">
-        <div class="info-user">
-          <div class="info-post">
-            <div class="photo-perfil-post">
-              <img src="img/perfilblack.png" alt="foto perfil de usuario">
+        <div class='post-publicated'>
+        <div class='info-user'>
+          <div class='info-post'>
+            <div class='photo-perfil-post'>
+              <img src='img/perfilblack.png' alt='foto perfil de usuario'>
             </div>
-            <div class="nameuser-date">
+            <div class='nameuser-date'>
               <span>${userName}</span> <br>
               <span>${post.data().datePost}</span>
             </div>
           </div>
-          <div class="btn-edit-delete">
-            <i class="bi bi-three-dots"></i>
+          <div class='btn-edit-delete'>
+            <i class='bi bi-three-dots'></i>
           </div>
         </div>
-        <div class="input-readonly">
-          <span id=${post.id} class="post-publicated cont-post" role="textbox">${post.data().post}</span>
+        <div class='input-readonly'>
+          <span id=${post.id} class='post-publicated cont-post' role='textbox'>${post.data().post}</span>
         </div>
       </div>
-      <div class="container-like-comment">
-        <div class="mando-img">
-          <i class="bi bi-joystick"></i>
-          <span class="span-text"> Me gusta </span>
+      <div class='container-like-comment'>
+        <div class='mando-img'>
+          <i class='bi bi-joystick'></i>
+          <span class='span-text'> Me gusta </span>
         </div>
-        <div class="comment-img">
-          <i class="bi bi-chat-dots"></i>
-          <span class="span-text"> Comentar </span>
-        </div>
-        <div>
-          <button type="button" class="btn-edit">Editar</button> 
+        <div class='comment-img'>
+          <i class='bi bi-chat-dots'></i>
+          <span class='span-text'> Comentar </span>
         </div>
         <div>
-          <button type="button" class="btn-delete">Eliminar</button> 
+          <button type='button' class='btn-edit'>Editar</button> 
         </div>
         <div>
-          <button type="button" class="btn-save">Guardar</button> 
+          <button type='button' class='btn-delete'>Eliminar</button> 
+        </div>
+        <div>
+          <button type='button' class='btn-save'>Guardar</button> 
         </div>
         </div>`;
 
@@ -145,9 +145,14 @@ export default () => {
 
         // EVENTO CLICK PARA ELIMINAR LOS POST
         btnDelete.addEventListener('click', () => {
-          deletePost(post.id).then(() => {
-            console.log('post eliminado ', post.id);
-          });
+          const opcion = window.confirm('¿Estás seguro que deseas eliminar el post?');
+          if (opcion === true) {
+            deletePost(post.id).then(() => {
+              console.log('post eliminado ', post.id);
+            });
+          } else {
+            console.log('Eliminación cancelada');
+          }
         });
         // EVENTOS DE CLICK PARA EDITAR LOS POST
         btnEdit.addEventListener('click', () => {
@@ -155,12 +160,17 @@ export default () => {
           contentPost.focus();
         });
         btnSave.addEventListener('click', () => {
-          editPost(post.id, {
-            post: contentPost.textContent,
-          }).then(() => {
-            contentPost.removeAttribute('contentEditable');
-            console.log('post editado');
-          });
+          const opcion = window.confirm('¿Estás seguro que deseas guardar los cambios?');
+          if (opcion === true) {
+            editPost(post.id, {
+              post: contentPost.textContent,
+            }).then(() => {
+              contentPost.removeAttribute('contentEditable');
+              console.log('post editado');
+            });
+          } else {
+            console.log('Edición cancelada');
+          }
         });
       });
     });
